requirements.txt
Django~=4.1
djangorestframework~=3.13.1
django-filter~=22.1
drf-yasg~=1.21.3
psycopg2-binary~=2.9.3